Changed checkpoint radius type

+++ backend/services/route-service/src/main/java/ru/nord/siwatch/backend/services/route/api/dto/CheckPointDto.java
@@ -9,7 +9,7 @@ import java.util.Date;
 @Setter
 public class CheckPointDto extends AbstractEntityDto {
 
-    private Integer radius;
+    private Double radius;
 
     private Long order;
